backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
azure-ai-voicelive>=1.0.0
azure-identity>=1.15.0
openai>=1.76.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0